[BaekJun/C++] 18111: Minecraft
This C++ program is designed to solve a problem related to terrain leveling, which is commonly encountered in competitive programming. The goal is to determine the minimum time required to make all terrain blocks at the same height and to identify the maximum possible height that can be achieved within that minimum time. The constraints involve the number of blocks available in the player's inventory and the costs associated with digging or stacking blocks.